+++ crates/core/vcs/tests/git_test.rs
@@ -36,7 +36,7 @@ mod file_hashing {
         let git = Git::load(&create_config("default"), sandbox.path()).unwrap();
 
         assert_eq!(
-            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], false)
+            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], false, None)
                 .await
                 .unwrap(),
             BTreeMap::from([
@@ -62,7 +62,7 @@ mod file_hashing {
         let git = Git::load(&create_config("default"), sandbox.path()).unwrap();
 
         assert_eq!(
-            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], false)
+            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], false, None)
                 .await
                 .unwrap(),
             BTreeMap::from([(
@@ -82,7 +82,7 @@ mod file_hashing {
         let git = Git::load(&create_config("default"), sandbox.path()).unwrap();
 
         assert_eq!(
-            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], true)
+            git.get_file_hashes(&string_vec!["existing.txt", "rename-me.txt"], true, None)
                 .await
                 .unwrap(),
             BTreeMap::from([
@@ -105,8 +105,11 @@ mod file_hashing {
 
         let git = Git::load(&create_config("default"), sandbox.path()).unwrap();
 
+        let tree = git.get_file_tree(".").await.unwrap();
+        let hashes = git.get_file_hashes(&tree, false, None).await.unwrap();
+
         assert_eq!(
-            git.get_file_tree_hashes(".").await.unwrap(),
+            hashes,
             BTreeMap::from([
                 (
                     ".gitignore".to_owned(),
@@ -136,9 +139,13 @@ mod file_hashing {
         let git = Git::load(&create_config("master"), sandbox.path()).unwrap();
 
         assert_eq!(
-            git.get_file_hashes(&string_vec!["foo", "bar", "dir/baz", "dir/qux"], false)
-                .await
-                .unwrap(),
+            git.get_file_hashes(
+                &string_vec!["foo", "bar", "dir/baz", "dir/qux"],
+                false,
+                None
+            )
+            .await
+            .unwrap(),
             BTreeMap::from([
                 (
                     "dir/qux".to_owned(),
@@ -160,21 +167,8 @@ mod file_hashing {
         let git = Git::load(&create_config("master"), sandbox.path()).unwrap();
 
         assert_eq!(
-            git.get_file_tree_hashes(".").await.unwrap(),
-            BTreeMap::from([
-                (
-                    ".gitignore".to_owned(),
-                    "589c59be54beff591804a008c972e76dea31d2d1".to_owned()
-                ),
-                (
-                    "dir/qux".to_owned(),
-                    "100b0dec8c53a40e4de7714b2c612dad5fad9985".to_owned()
-                ),
-                (
-                    "foo".to_owned(),
-                    "257cc5642cb1a054f08cc83f2d943e56fd3ebe99".to_owned()
-                )
-            ])
+            git.get_file_tree(".").await.unwrap(),
+            string_vec![".gitignore", "dir/qux", "foo",]
         );
     }
 
@@ -189,7 +183,8 @@ mod file_hashing {
             fs::write(sandbox.path().join(format!("file{}", i)), i.to_string()).unwrap();
         }
 
-        let hashes = git.get_file_tree_hashes(".").await.unwrap();
+        let tree = git.get_file_tree(".").await.unwrap();
+        let hashes = git.get_file_hashes(&tree, false, None).await.unwrap();
 
         assert!(hashes.len() >= 10000);
 
@@ -206,7 +201,7 @@ mod file_hashing {
         fs::create_dir_all(sandbox.path().join("dir")).unwrap();
 
         assert_eq!(
-            git.get_file_hashes(&string_vec!["dir"], false)
+            git.get_file_hashes(&string_vec!["dir"], false, None)
                 .await
                 .unwrap(),
             BTreeMap::new()